requirements.txt
aiogram==3.13.0
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv[cli]==1.0.0
aiosqlite==0.19.0
asyncpg==0.29.0
redis==5.0.1
openai==1.6.1
matplotlib==3.8.2
pandas==2.1.4
cryptography==41.0.7
apscheduler==3.11.0
aiohttp==3.9.5
pytest==8.4.2
# aiogram-tests==1.0.3  # Incompatible with aiogram 3.13.0
seaborn==0.13.2
sentry-sdk==2.39.0
factory-boy==3.3.3
pydantic>=2.4.1,<2.9